requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg==3.2.1
asyncpg==0.29.0
pydantic>=2.7.0,<3
apscheduler==3.10.4
httpx==0.27.2
aiosmtplib==3.0.2
python-dateutil==2.9.0.post0
tzdata>=2024.1

